login button move to login activity.
