+++ app/src/main/java/com/aqua/hoophelper/LoginActivityViewModel.kt
@@ -1,16 +1,26 @@
 package com.aqua.hoophelper
 
+import android.content.Context
+import android.content.Intent
+import android.util.Log
 import androidx.core.app.ActivityCompat
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.ViewModel
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 
-const val RC_SIGN_IN = 0
-
-class LoginViewModel : ViewModel() {
+class LoginActivityViewModel: ViewModel() {
 
     fun signIn(googleSignInClient: GoogleSignInClient, activity: FragmentActivity) {
         val signInIntent = googleSignInClient.signInIntent
         ActivityCompat.startActivityForResult(activity, signInIntent, RC_SIGN_IN, null)
     }
+
+    fun getStatus(context: Context, intent: Intent) {
+        Log.d("login","${User.account}")
+        if(User.account != null) {
+//            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+//            context.startActivity(intent)
+        }
+    }
+
 }
