--- app/src/main/java/com/aqua/hoophelper/LoginFragment.kt
@@ -1,45 +1,0 @@
-package com.aqua.hoophelper
-
-import androidx.lifecycle.ViewModelProvider
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import com.aqua.hoophelper.databinding.LoginFragmentBinding
-import com.aqua.hoophelper.profile.ProfileViewModel
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-
-class LoginFragment : Fragment() {
-
-    private val viewModel: LoginViewModel by lazy {
-        ViewModelProvider(this).get(LoginViewModel::class.java)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-
-        val binding: LoginFragmentBinding =
-            DataBindingUtil.inflate(inflater, R.layout.login_fragment, container, false)
-
-        // Configure Google Sign In
-        val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-            .requestIdToken(getString(R.string.default_web_client_id))
-            .requestEmail()
-            .build()
-
-        // getting the value of gso inside the GoogleSignInClient
-        val googleSignInClient = GoogleSignIn.getClient(requireActivity(), gso)
-
-        binding.googleSignIn.setOnClickListener {
-            viewModel.signIn(googleSignInClient,requireActivity())
-
-        }
-
-        return binding.root
-    }
-}
