+++ app/src/main/java/com/aqua/hoophelper/MainActivity.kt
@@ -7,6 +7,7 @@ import android.os.Bundle
 import android.util.Log
 import android.view.KeyEvent
 import android.view.View
+import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.findNavController
@@ -30,9 +31,6 @@ import com.google.firebase.firestore.Query
 
 
 class MainActivity : AppCompatActivity() {
-
-    private lateinit var auth: FirebaseAuth
-
 
     private val viewModel: MainActivityViewModel by lazy {
         ViewModelProvider(this).get(MainActivityViewModel::class.java)
@@ -73,24 +71,6 @@ class MainActivity : AppCompatActivity() {
                 }
         }
 
-        // login or not
-        viewModel.loginState.observe(this){
-            if (it) {
-                Log.d("loginn","yes")
-                binding.toolbar.visibility = View.VISIBLE
-                binding.bottomBar.visibility = View.VISIBLE
-                binding.fab.visibility = View.VISIBLE
-                binding.appBar.visibility = View.VISIBLE
-                navHostFragment.navigate(NavigationDirections.navToHome())
-            } else {
-                Log.d("loginn","no")
-                binding.toolbar.visibility = View.GONE
-                binding.bottomBar.visibility = View.GONE
-                binding.fab.visibility = View.GONE
-                binding.appBar.visibility = View.GONE
-                navHostFragment.navigate(NavigationDirections.navToLogin())
-            }
-        }
 
         navHostFragment.addOnDestinationChangedListener { _, destination, _ ->
             when(destination.id) {
@@ -114,68 +94,4 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-
-    override fun onStart() {
-        super.onStart()
-        // Initialize Firebase Auth
-        auth = Firebase.auth
-        // Check if user is signed in (non-null) and update UI accordingly.
-        val currentUser = auth.currentUser
-        User.account = currentUser
-        Log.d("currentUser","${currentUser?.email}")
-        viewModel.getUserInfo()
-        viewModel.checkLogin(User.account)
-        updateUI(currentUser)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-
-        if (requestCode == RC_SIGN_IN) {
-            val task = GoogleSignIn.getSignedInAccountFromIntent(data)
-            try {
-                // Google Sign In was successful, authenticate with Firebase
-                val account = task.getResult(ApiException::class.java)
-                Log.d("login", "firebaseAuthWithGoogle:" + account.email)
-                viewModel.checkLogin(account)
-//                Log.d("loginn","${viewModel.loginState.value}")
-                firebaseAuthWithGoogle(account.idToken!!)
-            } catch (e: ApiException) {
-                // Google Sign In failed, update UI appropriately
-                Log.w("login", "Google sign in failed", e)
-            }
-        }
-    }
-
-    private fun firebaseAuthWithGoogle(idToken: String) {
-        val credential = GoogleAuthProvider.getCredential(idToken, null)
-        auth.signInWithCredential(credential)
-            .addOnCompleteListener(this) { task ->
-                if (task.isSuccessful) {
-                    // Sign in success, update UI with the signed-in user's information
-                    Log.d("login", "signInWithCredential:success")
-                    val user = auth.currentUser
-                    updateUI(user)
-                } else {
-                    // If sign in fails, display a message to the user.
-                    Log.w("login", "signInWithCredential:failure", task.exception)
-                    updateUI(null)
-                }
-            }
-    }
-
-    private fun updateUI(user: FirebaseUser?) {
-
-    }
-}
-
-object User {
-    var account: FirebaseUser? = null
-    var teamId = ""
-    var teamMembers = MutableLiveData(listOf<Player>())
-}
-
-object HoopInfo {
-    var spinnerSelectedTeamId = ""
-    var matchId = ""
 }
