+++ app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/basil_bg"
-    tools:context=".LoginFragment">
+    tools:context=".LoginActivity">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/google_sign_in"
