+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">HoopHelper</string>
 
-<!--  team List  -->
+    <!--  team List  -->
     <string-array name="Team">
         <item>Jazz</item>
         <item>Sun</item>
